+++ scrolls/mysql.rb
@@ -1,14 +1,8 @@
 gem "mysql2"
 
-# TODO what about Windows?
-
-# if scroll?("eycloud")
-#   mysql_versions = [
-#     ["MySQL 5.0", "mysql_50"],
-#     ["MySQL 5.5 (beta)", "mysql_55"]
-#   ]
-#   @mysql_stack = multiple_choice("Create app to which Engine Yard Cloud account?", mysql_versions)
-# end
+if scroll?("eycloud")
+  @db_stack = "mysql"
+end
 
 after_bundler do
   rake "db:create:all" if config['auto_create']
@@ -42,4 +36,4 @@ args: -d mysql
 config:
   - auto_create:
       type: boolean
-      prompt: "Create MySQL database with default configuration?"
+      prompt: "Create local MySQL databases with default configuration?"