+++ scrolls/postgresql.rb
@@ -1,26 +1,25 @@
 gem "pg"
 
 gsub_file "config/database.yml", /username: .*/, "username: #{config['pg_username']}"
-gsub_file "config/database.yml", /password: .*/, "username: #{config['pg_password']}"
+gsub_file "config/database.yml", /password: .*/, "password: #{config['pg_password']}"
 
-# if scroll?("eycloud")
-#   @db_stack = "postgresql_91"
-# end
+if scroll?("eycloud")
+  @db_stack = "postgresql"
+end
 
 after_bundler do
   rake "db:create:all" if config['auto_create']
   
-  if config['populate_rake_task']
-    sample_rake = <<-RUBY
-require './config/environment'
+  rakefile("sample.rake") do
+<<-RUBY
 namespace :db do
   desc "Populate the database with sample data"
-  task :sample do
+  task :sample => :environment do
+
   end
   task :populate => :sample
 end
 RUBY
-    File.open("lib/tasks/sample.rake", 'w') {|f| f.write(sample_rake)}
   end
 end
 
@@ -40,11 +39,7 @@ args: -d postgresql
 config:
   - auto_create:
       type: boolean
-      prompt: "Create PostgreSQL database with default configuration?"
-
-  - populate_rake_task:
-      type: boolean
-      prompt: "Add db:sample rake task?"
+      prompt: "Create local PostgreSQL databases with default configuration?"
 
   - pg_username:
       type: string